feat!: extract schema inference to SchemaOptions
To enable customization of the schema, and ensure the marshaling and
unmarshaling is consistent with those options, this commit updates the
API of the package to introduce the SchemaOptions type, which is
included in both the MarshalOptions and UnmarshalOptions.

No SchemaOptions will be implemented until full support for the
"default" schema options is implemented.
