+++ encoding/protobq/encode_test.go
@@ -84,46 +84,3 @@ func TestMarshalOptions_Marshal(t *testing.T) {
 		})
 	}
 }
-
-func TestMarshalOptions_InferSchema(t *testing.T) {
-	for _, tt := range []struct {
-		name     string
-		msg      proto.Message
-		opt      MarshalOptions
-		expected bigquery.Schema
-	}{
-		{
-			name: "library.Book",
-			msg:  &library.Book{},
-			expected: bigquery.Schema{
-				{Name: "name", Type: bigquery.StringFieldType},
-				{Name: "author", Type: bigquery.StringFieldType},
-				{Name: "title", Type: bigquery.StringFieldType},
-				{Name: "read", Type: bigquery.BooleanFieldType},
-			},
-		},
-		{
-			name: "library.UpdateBookRequest",
-			msg:  &library.UpdateBookRequest{},
-			expected: bigquery.Schema{
-				{Name: "name", Type: bigquery.StringFieldType},
-				{
-					Name: "book",
-					Type: bigquery.RecordFieldType,
-					Schema: bigquery.Schema{
-						{Name: "name", Type: bigquery.StringFieldType},
-						{Name: "author", Type: bigquery.StringFieldType},
-						{Name: "title", Type: bigquery.StringFieldType},
-						{Name: "read", Type: bigquery.BooleanFieldType},
-					},
-				},
-			},
-		},
-	} {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			actual := tt.opt.InferSchema(tt.msg)
-			assert.DeepEqual(t, tt.expected, actual)
-		})
-	}
-}
