+++ encoding/protobq/schema_test.go
@@ -1,0 +1,53 @@
+package protobq
+
+import (
+	"testing"
+
+	"cloud.google.com/go/bigquery"
+	"google.golang.org/genproto/googleapis/example/library/v1"
+	"google.golang.org/protobuf/proto"
+	"gotest.tools/v3/assert"
+)
+
+func TestSchemaOptions_InferSchema(t *testing.T) {
+	for _, tt := range []struct {
+		name     string
+		msg      proto.Message
+		opt      SchemaOptions
+		expected bigquery.Schema
+	}{
+		{
+			name: "library.Book",
+			msg:  &library.Book{},
+			expected: bigquery.Schema{
+				{Name: "name", Type: bigquery.StringFieldType},
+				{Name: "author", Type: bigquery.StringFieldType},
+				{Name: "title", Type: bigquery.StringFieldType},
+				{Name: "read", Type: bigquery.BooleanFieldType},
+			},
+		},
+		{
+			name: "library.UpdateBookRequest",
+			msg:  &library.UpdateBookRequest{},
+			expected: bigquery.Schema{
+				{Name: "name", Type: bigquery.StringFieldType},
+				{
+					Name: "book",
+					Type: bigquery.RecordFieldType,
+					Schema: bigquery.Schema{
+						{Name: "name", Type: bigquery.StringFieldType},
+						{Name: "author", Type: bigquery.StringFieldType},
+						{Name: "title", Type: bigquery.StringFieldType},
+						{Name: "read", Type: bigquery.BooleanFieldType},
+					},
+				},
+			},
+		},
+	} {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			actual := tt.opt.InferSchema(tt.msg)
+			assert.DeepEqual(t, tt.expected, actual)
+		})
+	}
+}
